fix: handle reconnect
After updating the firmware on my PPK2 to v4.2.0 I noticed that
connecting, disconnected then trying to reconnect failed.
It seems like the updated firmware has some 0x80 byte that gets
send and as a result an exception gets thrown.

This commit allows those bytes to be replaced, thus the check is still valid
but we do not get this exception.